+++ src/navigation.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-
 import Onboarding from './screens/onboarding/Onboarding';
 import EnableNotifications from './screens/onboarding/EnableNotifications';
 import WalletBackup from './screens/onboarding/WalletBackup';
@@ -13,73 +11,9 @@ import Legal1And2 from './screens/Legal1And2';
 import PassCodeEntry1 from './screens/PassCodeEntry1';
 import CodeEntry from './screens/CodeEntry';
 
-import * as colors from './constants/colors';
-import { Image } from 'react-native';
-import TabIcon from './components/common/TabIcon';
-import { useTheme } from './contexts/theme';
+import Tabs from './components/Tabs/BottomNavigationTabs';
 
 const { Group, Navigator, Screen } = createNativeStackNavigator();
-const Tab = createBottomTabNavigator();
-
-const Tabs = () => {
-  const theme = useTheme();
-  return (
-    <Tab.Navigator
-      initialRouteName="Home"
-      screenOptions={{
-        headerShown: false,
-        tabBarActiveTintColor: colors.blue,
-        tabBarInactiveTintColor: colors.lightGrey,
-        tabBarLabelStyle: {
-          fontFamily: 'Sf'
-        },
-        tabBarStyle: {
-          backgroundColor: theme.mode == 'dark' ? colors.appBGDark : colors.white,
-          borderTopWidth: 0
-        },
-        indicatorStyle: {
-          height: 0,
-          backgroundColor: 'pink'
-        }
-      }}
-    >
-      <Tab.Screen
-        name="Wallet"
-        component={Legal1And2}
-        options={{
-          tabBarIcon: ({ color }) => <TabIcon size={25} name="wallet-outline" color={color} />
-        }}
-      />
-      <Tab.Screen
-        name="Transactions"
-        component={SecretPhrase}
-        options={{
-          tabBarIcon: ({ color }) => <TabIcon size={25} name="grid-outline" color={color} />
-        }}
-      />
-      <Tab.Screen
-        name="Dex"
-        component={VerifySecretPhrase}
-        options={{
-          tabBarIcon: ({ color }) => {
-            if (color == '#105DCF') {
-              return <Image source={require('../assets/icons/dex.png')} style={{ height: 25, width: 25 }} />;
-            } else {
-              return <Image source={require('../assets/icons/dex-inactive.png')} style={{ height: 25, width: 25 }} />;
-            }
-          }
-        }}
-      />
-      <Tab.Screen
-        name="Settings"
-        component={Settings}
-        options={{
-          tabBarIcon: ({ color }) => <TabIcon size={25} name="settings-outline" color={color} />
-        }}
-      />
-    </Tab.Navigator>
-  );
-};
 
 const Navigation = () => {
   return (
